Separated back-end machine learning script into Train and Run Scripts
The machLearn-Train.py is run once to train the models and save the trained model states, and then the machLearn-Run calls these saved states every time. For example machLearn-Run.py will be the file called by the actual web app.
